+++ src/texture.hpp
@@ -10,11 +10,11 @@ class TextureHandle {
 public:
     TextureHandle();
 
-    TextureHandle(uint32_t value);
+    TextureHandle(uint64_t value);
 
-    void set(uint32_t value);
+    void set(uint64_t value);
 
-    [[nodiscard]] uint32_t value() const;
+    [[nodiscard]] uint64_t value() const;
 
     [[nodiscard]] bool operator==(const TextureHandle& other) const;
 
@@ -22,7 +22,7 @@ public:
 
 private:
     bool m_initialized = false;
-    uint32_t m_value;
+    uint64_t m_value;
 };
 
 class Renderer;
@@ -64,7 +64,7 @@ template <>
 struct hash<mve::TextureHandle> {
     std::size_t operator()(const mve::TextureHandle& handle) const
     {
-        return hash<uint32_t>()(handle.value());
+        return hash<uint64_t>()(handle.value());
     }
 };
