+++ src/uniform_buffer.hpp
@@ -14,11 +14,11 @@ class UniformBufferHandle {
 public:
     UniformBufferHandle();
 
-    UniformBufferHandle(uint32_t value);
+    UniformBufferHandle(uint64_t value);
 
-    void set(uint32_t value);
+    void set(uint64_t value);
 
-    [[nodiscard]] uint32_t value() const;
+    [[nodiscard]] uint64_t value() const;
 
     [[nodiscard]] bool operator==(const UniformBufferHandle& other) const;
 
@@ -26,7 +26,7 @@ public:
 
 private:
     bool m_initialized = false;
-    uint32_t m_value;
+    uint64_t m_value;
 };
 
 class Renderer;
@@ -84,7 +84,7 @@ template <>
 struct hash<mve::UniformBufferHandle> {
     std::size_t operator()(const mve::UniformBufferHandle& handle) const
     {
-        return hash<uint32_t>()(handle.value());
+        return hash<uint64_t>()(handle.value());
     }
 };
